refactor map helper functions into utils
