+++ packages/site/src/routes/index.svelte
@@ -28,7 +28,7 @@
   import ToggleStyle from '@living-dictionaries/parts/src/lib/maps/mapbox/controls/ToggleStyle.svelte';
   import NavigationControl from '@living-dictionaries/parts/src/lib/maps/mapbox/controls/NavigationControl.svelte';
   import CustomControl from '@living-dictionaries/parts/src/lib/maps/mapbox/controls/CustomControl.svelte';
-  import { getTimeZoneLongitude } from '@living-dictionaries/parts/src/lib/maps/getTimeZoneLongitude';
+  import { getTimeZoneLongitude } from '@living-dictionaries/parts/src/lib/maps/utils/getTimeZoneLongitude';
   import DictionaryPoints from '$lib/components/home/DictionaryPoints.svelte';
   import Search from '$lib/components/home/Search.svelte';
   import Header from '$lib/components/shell/Header.svelte';
